requirements.txt
python-nmap
jinja2
prettytable
